Allow offset to be floats

+++ app/types.py
@@ -7,5 +7,5 @@ from PIL.ImageFont import FreeTypeFont as FontType  # noqa
 Box = tuple[int, int, int, int]
 Dimensions = tuple[int, int]
 Point = tuple[int, int]
-Offset = tuple[int, int]
+Offset = tuple[float, float]
 Align = Union[Literal["left"], Literal["center"], Literal["right"]]
